+++ CHANGELOG.md
@@ -16,9 +16,8 @@ Initial working version
  *  Support for dynamics and accents on individual notes
      *  Accents currently supported: Accent, Marcato, Staccato, Staccatissimo, Tenuto, Slurs, Ties
  *  Specify musical parts with time, key, tempo, and dynamic changes
-     *  **NOTE**: individual parts do not currently support polyphony
  *  Support for customizable instruments with adjustable dynamic ranges and note envelopes
      *  Synthesizers: support sine, square, triangle, and sawtooth waves
      *  Samplers: support for playing back samples from WAV files
- *  Polyphonics (via multiple parts. Each part is monophonic)
+ *  Polyphonics (both via multiple parts and by chording within parts)
  *  Writing to WAV files (in stereo or mono)